fix: rutas públicas accesibles y argon2 fuera del bundle cliente
- auth.config: agrega /register, /login, /forgot-password, /reset-password
  y /map como rutas públicas en authorized callback
- Separa validatePasswordStrength a password-validation.ts (sin imports Node)
  para que register-form y reset-password-form no arrastren argon2 al bundle
- password.ts re-exporta validatePasswordStrength para compatibilidad
